Rename `mesh._base` to `mesh._line_rectangle_cube`
